Tweak the lead indicator code to not just work with primaries -- now, dumbfire secondaries show the correct position. Also, make it work when the target is stationary but "$use additive weapon velocity" is set. Fix a couple typos and warnings.

+++ fs2_open/code/ai/ai_profiles.h
@@ -116,8 +116,8 @@ typedef struct ai_profile_t {
 	float ship_fire_delay_scale_friendly[NUM_SKILL_LEVELS];
 
 	//	Maximum turrets of one ship allowed to shoot the same target
-	float max_turret_ownage_target[NUM_SKILL_LEVELS];
-	float max_turret_ownage_player[NUM_SKILL_LEVELS];
+	int max_turret_ownage_target[NUM_SKILL_LEVELS];
+	int max_turret_ownage_player[NUM_SKILL_LEVELS];
 
 } ai_profile_t;
